I have removed the TAXII header from the TAXII HTTP Protocol Binding Spec. This is updated in the most recent commit to the GitHub repository. You can also find a reference to the issue here: https://github.com/TAXIIProject/TAXII-Specifications/issues/14.
The essence of the change is that the X-TAXII-Accept header is no longer a
concept in the TAXII HTTP Protocol binding. Software that provides a TAXII
Service can assume that the client accepts the format specified by the
X-TAXII-Content-Type.

A quick summary of locations where text was changed:

Removal of X-TAXII-Accept:
-	1.2.5, Media Range
-	3 - TAXII Version-Binding Type
-	Table 1 (HTTP Headers)
-	4.1.1 Accept
-	4.1.3 - X-TAXII-Accept
-	5.2 - Request Headers
-	6.1 - Response Headers
-	7 - HTTP Status Codes and TAXII Error Messages

Added the following text to the X-TAXII-Content-Type explaining how to
infer an appropriate response format:
When responding to a request, the server response MUST comply with the
format specified by the request's X-TAXII-Content-Type header field. If
the server is unable to respond with an appropriate format, it must send
an error.

Other changes:
1. Made the line spacing uniform in section 5.1
2. Section 5.1.4 - TAXII Poll Request had conflicting wording. Changed the
wording to reflect that a TAXII Poll Request is sent via HTTP Post, as
intended.
3. Updated the title/watermark to draft2. I will discuss this in a
separate email.